:sparkles: Create the sidebar component
create the side bar component with for desktop
